AI_Fellow_webapp/requirements.txt
Django==3.0.2
requests==2.18.4
flickrapi==2.4.0
Pillow==7.0.0
